+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Values with a different nesting could have the same universal hash.
 - Values containing `bytes` or `str` of an arbitrary length could have the same universal hash.
 - Universal hashing raised `RecursionError` for deeply nested values and circular references.
+- Universal hashing raised `TypeError` for mappings and sets with uncomparable keys or items.
 
 ### Changed